Apps/gen.ia creators/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pillow==10.2.0
anthropic>=0.40.0
python-multipart==0.0.9
pydantic==2.6.3
requests==2.31.0
python-dotenv==1.0.1
